refact: clean vqueue file tree
