+++ vqueue/src/lib.rs
@@ -9,7 +9,6 @@
 #![warn(clippy::nursery)]
 #![warn(clippy::cargo)]
 //
-#![allow(clippy::doc_markdown)]
 #![allow(clippy::multiple_crate_versions)]
 
 mod args;
@@ -19,16 +18,3 @@ mod model;
 pub use args::{Args, Commands, MessageCommand, MessageShowFormat};
 pub use command::execute;
 pub use model::{QueueContent, QueueEntry};
-
-/// Generate the list of lifetime
-#[must_use]
-pub fn lifetimes() -> Vec<u64> {
-    (0..9)
-        .into_iter()
-        .scan(5, |state, _| {
-            let out = *state;
-            *state *= 2;
-            Some(out)
-        })
-        .collect()
-}
